requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pandas==2.2.2
requests==2.32.3
